removed core and draw engine code, updated net version
