+++ src/AttractorViewer/AttractorViewer.csproj
@@ -2,21 +2,19 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net4.8</TargetFramework>
+    <TargetFramework>net6.0-windows</TargetFramework>
+    <LangVersion>10</LangVersion>
     <UseWPF>true</UseWPF>
     <ApplicationIcon>app.ico</ApplicationIcon>
     <PlatformTarget>AnyCPU</PlatformTarget>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="ChaosSoft.Core" Version="1.0.0" />
+  </ItemGroup>
+
+    <ItemGroup>
     <Content Include="app.ico" />
   </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ChaosSoft.Core\ChaosSoft.Core.csproj" />
-    <ProjectReference Include="..\ChaosSoft.DrawEngine\ChaosSoft.DrawEngine.csproj" />
-  </ItemGroup>
-
-
 
 </Project>
