+++ src/AttractorViewer/ColorMaps/IColorMap.cs
@@ -1,0 +1,14 @@
+﻿using System.Drawing;
+
+namespace AttractorViewer.ColorMaps;
+
+public enum ColorMap
+{
+    Orange,
+    Parula
+}
+
+public interface IColorMap
+{
+    Color GetColor(double value);
+}
