+++ src/AttractorViewer/SegmentsCreator.cs
@@ -1,104 +1,103 @@
 ﻿using System;
 using System.Windows.Media.Media3D;
 
-namespace AttractorViewer
+namespace AttractorViewer;
+
+public class SegmentsCreator
 {
-    public class SegmentsCreator
+
+    public SegmentsCreator()
     {
 
-        public SegmentsCreator()
-        {
+    }
 
-        }
+    public GeometryModel3D CreateModel(Material material, Point3D point1, Point3D point2, double thickness)
+    {
+        var mesh = new MeshGeometry3D();
 
-        public GeometryModel3D CreateModel(Material material, Point3D point1, Point3D point2, double thickness)
-        {
-            var mesh = new MeshGeometry3D();
+        // Find an up vector that is not colinear with the segment.
+        // Start with a vector parallel to the Y axis.
+        Vector3D up = new Vector3D(0, 1, 0);
 
-            // Find an up vector that is not colinear with the segment.
-            // Start with a vector parallel to the Y axis.
-            Vector3D up = new Vector3D(0, 1, 0);
+        // If the segment and up vector point in more or less the
+        // same direction, use an up vector parallel to the X axis.
+        Vector3D segment = point2 - point1;
+        segment.Normalize();
+        if (Math.Abs(Vector3D.DotProduct(up, segment)) > 0.9)
+            up = new Vector3D(1, 0, 0);
 
-            // If the segment and up vector point in more or less the
-            // same direction, use an up vector parallel to the X axis.
-            Vector3D segment = point2 - point1;
-            segment.Normalize();
-            if (Math.Abs(Vector3D.DotProduct(up, segment)) > 0.9)
-                up = new Vector3D(1, 0, 0);
+        // Add the segment.
+        AddSegment(mesh, point1, point2, up, thickness);
 
-            // Add the segment.
-            AddSegment(mesh, point1, point2, up, thickness);
+        return new GeometryModel3D(mesh, material);
+    }
 
-            return new GeometryModel3D(mesh, material);
-        }
+    public void AddSegment(MeshGeometry3D mesh, Point3D point1, Point3D point2, Vector3D up, double thickness)
+    {
+        // Get the segment's vector.
+        Vector3D v = point2 - point1;
 
-        public void AddSegment(MeshGeometry3D mesh, Point3D point1, Point3D point2, Vector3D up, double thickness)
-        {
-            // Get the segment's vector.
-            Vector3D v = point2 - point1;
+        // Get the scaled up vector.
+        Vector3D n1 = ScaleVector(up, thickness / 2.0);
 
-            // Get the scaled up vector.
-            Vector3D n1 = ScaleVector(up, thickness / 2.0);
+        // Get another scaled perpendicular vector.
+        Vector3D n2 = Vector3D.CrossProduct(v, n1);
+        n2 = ScaleVector(n2, thickness / 2.0);
 
-            // Get another scaled perpendicular vector.
-            Vector3D n2 = Vector3D.CrossProduct(v, n1);
-            n2 = ScaleVector(n2, thickness / 2.0);
+        // Make a skinny box.
+        // p1pm means point1 PLUS n1 MINUS n2.
+        Point3D p1pp = point1 + n1 + n2;
+        Point3D p1mp = point1 - n1 + n2;
+        Point3D p1pm = point1 + n1 - n2;
+        Point3D p1mm = point1 - n1 - n2;
+        Point3D p2pp = point2 + n1 + n2;
+        Point3D p2mp = point2 - n1 + n2;
+        Point3D p2pm = point2 + n1 - n2;
+        Point3D p2mm = point2 - n1 - n2;
 
-            // Make a skinny box.
-            // p1pm means point1 PLUS n1 MINUS n2.
-            Point3D p1pp = point1 + n1 + n2;
-            Point3D p1mp = point1 - n1 + n2;
-            Point3D p1pm = point1 + n1 - n2;
-            Point3D p1mm = point1 - n1 - n2;
-            Point3D p2pp = point2 + n1 + n2;
-            Point3D p2mp = point2 - n1 + n2;
-            Point3D p2pm = point2 + n1 - n2;
-            Point3D p2mm = point2 - n1 - n2;
+        // Sides.
+        AddTriangle(mesh, p1pp, p1mp, p2mp);
+        AddTriangle(mesh, p1pp, p2mp, p2pp);
 
-            // Sides.
-            AddTriangle(mesh, p1pp, p1mp, p2mp);
-            AddTriangle(mesh, p1pp, p2mp, p2pp);
+        AddTriangle(mesh, p1pp, p2pp, p2pm);
+        AddTriangle(mesh, p1pp, p2pm, p1pm);
 
-            AddTriangle(mesh, p1pp, p2pp, p2pm);
-            AddTriangle(mesh, p1pp, p2pm, p1pm);
+        AddTriangle(mesh, p1pm, p2pm, p2mm);
+        AddTriangle(mesh, p1pm, p2mm, p1mm);
 
-            AddTriangle(mesh, p1pm, p2pm, p2mm);
-            AddTriangle(mesh, p1pm, p2mm, p1mm);
+        AddTriangle(mesh, p1mm, p2mm, p2mp);
+        AddTriangle(mesh, p1mm, p2mp, p1mp);
 
-            AddTriangle(mesh, p1mm, p2mm, p2mp);
-            AddTriangle(mesh, p1mm, p2mp, p1mp);
+        // Ends.
+        AddTriangle(mesh, p1pp, p1pm, p1mm);
+        AddTriangle(mesh, p1pp, p1mm, p1mp);
 
-            // Ends.
-            AddTriangle(mesh, p1pp, p1pm, p1mm);
-            AddTriangle(mesh, p1pp, p1mm, p1mp);
+        AddTriangle(mesh, p2pp, p2mp, p2mm);
+        AddTriangle(mesh, p2pp, p2mm, p2pm);
+    }
 
-            AddTriangle(mesh, p2pp, p2mp, p2mm);
-            AddTriangle(mesh, p2pp, p2mm, p2pm);
-        }
+    // Add a triangle to the indicated mesh.
+    // Do not reuse points so triangles don't share normals.
+    private void AddTriangle(MeshGeometry3D mesh, Point3D point1, Point3D point2, Point3D point3)
+    {
+        // Create the points.
+        int index1 = mesh.Positions.Count;
+        mesh.Positions.Add(point1);
+        mesh.Positions.Add(point2);
+        mesh.Positions.Add(point3);
 
-        // Add a triangle to the indicated mesh.
-        // Do not reuse points so triangles don't share normals.
-        private void AddTriangle(MeshGeometry3D mesh, Point3D point1, Point3D point2, Point3D point3)
-        {
-            // Create the points.
-            int index1 = mesh.Positions.Count;
-            mesh.Positions.Add(point1);
-            mesh.Positions.Add(point2);
-            mesh.Positions.Add(point3);
+        // Create the triangle.
+        mesh.TriangleIndices.Add(index1++);
+        mesh.TriangleIndices.Add(index1++);
+        mesh.TriangleIndices.Add(index1);
+    }
 
-            // Create the triangle.
-            mesh.TriangleIndices.Add(index1++);
-            mesh.TriangleIndices.Add(index1++);
-            mesh.TriangleIndices.Add(index1);
-        }
-
-        private Vector3D ScaleVector(Vector3D vector, double length)
-        {
-            double scale = length / vector.Length;
-            return new Vector3D(
-                vector.X * scale,
-                vector.Y * scale,
-                vector.Z * scale);
-        }
+    private Vector3D ScaleVector(Vector3D vector, double length)
+    {
+        double scale = length / vector.Length;
+        return new Vector3D(
+            vector.X * scale,
+            vector.Y * scale,
+            vector.Z * scale);
     }
 }
