--- src/ChaosSoft.Core/ChaosSoft.Core.nuspec
@@ -1,34 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>ChaosSoft.Core</id>
-    <version>1.0.0</version>
-    <title>ChaosSoft.Core</title>
-    <authors>Vitaliy Dobriyan</authors>
-    <license type="expression">Apache-2.0</license>
-    <copyright>Copyright (c) Vitaliy Dobriyan 2022</copyright>
-    <projectUrl>https://bitbucket.org/dobriyanchik/chaossoft</projectUrl>
-    <repository type="git" url="https://bitbucket.org/dobriyanchik/chaossoft"/>
-    <icon>images\icon.png</icon>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>
-      This package includes base data IO, Timeseries related objects, numerical methods.
-    </description>
-    <releaseNotes>First release: base data IO, Timeseries related objects, numerical methods.</releaseNotes>
-    <tags>math numerical-methods lyapunov</tags>
-    <dependencies>
-      <group targetFramework="net452"/>
-      <group targetFramework="net5.0"/>
-      <group targetFramework="netstandard2.0"/>
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="..\..\icon.png" target="images\" />
-    <file src="bin\Release\net452\chaossoft.core.dll" target="lib\net452" />
-    <file src="bin\Release\net452\chaossoft.core.xml" target="lib\net452" />
-    <file src="bin\Release\netstandard2.0\chaossoft.core.dll" target="lib\netstandard2.0" />
-    <file src="bin\Release\netstandard2.0\chaossoft.core.xml" target="lib\netstandard2.0" />
-    <file src="bin\Release\net5.0\chaossoft.core.dll" target="lib\net5.0" />
-    <file src="bin\Release\net5.0\chaossoft.core.xml" target="lib\net5.0" />
-  </files>
-</package>
