--- src/ChaosSoft.DrawEngine/ColorMaps/IColorMap.cs
@@ -1,15 +1,0 @@
-﻿using System.Drawing;
-
-namespace ChaosSoft.DrawEngine.ColorMaps
-{
-    public enum ColorMap
-    {
-        Orange,
-        Parula
-    }
-
-    public interface IColorMap
-    {
-        Color GetColor(double value);
-    }
-}
