+++ src/ChaosSoft.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32526.322
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChaosSoft.Core", "ChaosSoft.Core\ChaosSoft.Core.csproj", "{EDA2C4F0-1D19-4FC8-A608-5B0607EEA67A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AttractorViewer", "AttractorViewer\AttractorViewer.csproj", "{6B475299-098A-4934-B27A-064D75AD5C6A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChaosSoft.DrawEngine", "ChaosSoft.DrawEngine\ChaosSoft.DrawEngine.csproj", "{7C94C80E-20C5-4707-8AB1-B37BEA92AA11}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1F0C3803-792C-466E-A68A-3864F3AF663D}"
 	ProjectSection(SolutionItems) = preProject
@@ -20,18 +16,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EDA2C4F0-1D19-4FC8-A608-5B0607EEA67A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EDA2C4F0-1D19-4FC8-A608-5B0607EEA67A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EDA2C4F0-1D19-4FC8-A608-5B0607EEA67A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EDA2C4F0-1D19-4FC8-A608-5B0607EEA67A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6B475299-098A-4934-B27A-064D75AD5C6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6B475299-098A-4934-B27A-064D75AD5C6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6B475299-098A-4934-B27A-064D75AD5C6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6B475299-098A-4934-B27A-064D75AD5C6A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7C94C80E-20C5-4707-8AB1-B37BEA92AA11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C94C80E-20C5-4707-8AB1-B37BEA92AA11}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C94C80E-20C5-4707-8AB1-B37BEA92AA11}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C94C80E-20C5-4707-8AB1-B37BEA92AA11}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
